Disc Projectile texture ALMOST working (need to add charge position variation) Disc Behaviour with jukeboxes working
TODO : need to add base damage and support for trowing it

+++ src/client/java/me/gv0id/arbalests/client/render/entity/state/MusicDiscEntityState.java
@@ -9,7 +9,7 @@ import net.minecraft.util.math.Vec3d;
 @Environment(EnvType.CLIENT)
 public class MusicDiscEntityState extends EntityRenderState {
     public final ItemRenderState itemRenderState = new ItemRenderState();
-    public int rotation;
+    public float rotation;
     public boolean glow;
     public float pitch;
     public float yaw;
